# docs/implementation/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/contributions/{id}": {
      "get": {
        "operationId": "ContributionsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
    "/api/v1/contributions/{id}/revisions": {
      "post": {
        "operationId": "ContributionsController_revise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
    "/api/v1/contributions/{id}/consent": {
      "post": {
        "operationId": "ContributionsController_coauthor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
    "/api/v1/contributions/{id}/submit": {
      "post": {
        "operationId": "ContributionsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
    "/api/v1/contributions/{id}/withdraw-submission": {
      "post": {
        "operationId": "ContributionsController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
    "/api/v1/editor/contributions/{id}/assign": {
      "post": {
        "operationId": "ContributionsController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
    "/api/v1/review-assignments/{id}/decisions": {
      "post": {
        "operationId": "ContributionsController_decide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Contributions"
        ]
      }
    },
...
  },
  "info": {
    "title": "HumanScope REST API",
    "description": "Foundation route inventory. Body validation is defined by Zod contracts; response schemas and full OpenAPI coverage remain a release gate.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hs_session"
      }
    },
...
  }
}